Fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@44798 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/userport/userport_wic64.c
@@ -1367,7 +1367,7 @@ static size_t post_write_func(char *buffer, size_t size, size_t nitems, void *us
     }
     memcpy(post_data + post_data_rcvd, buffer, ret);
     post_data_rcvd += ret;
-    _hexdump(CONS_COL_NO, 2, buffer, ret);
+    _hexdump(CONS_COL_NO, 2, buffer, (int)ret);
   out:
     return ret;
 }
